Use trait for CompoundDeviceIdentifier.
Allow to better create "remote CDIs", by separating the implementation
and the interface for the rest of rice code for the CDI.

Also remove previously introduced `from_raw_parts`.
